avoid code duplication with jitdump_linux
